dev_requirements.txt
pytest>=2.7
pytest-cov
pytest-randomly
pytest-mock
pylint
xmltodict
autopep8
flake8
